requirements.txt
streamlit
groq
chromadb
sentence-transformers
streamlit-authenticator
pyyaml
python-dotenv